Use build_command module
