+++ crates/sandbox/src/lib.rs
@@ -18,8 +18,10 @@
 //! In order to ease testing, the service is parameterized by a compile
 //! strategy. This allows easy mocking.
 
-pub mod example_code;
+mod build_command;
+mod example_code;
 
+use crate::build_command::build_compile_command;
 use serde::{
     Deserialize,
     Serialize,
@@ -135,15 +137,7 @@ pub enum CompilationResult {
 // CONSTANTS
 // -------------------------------------------------------------------------------------------------
 
-const DOCKER_PROCESS_TIMEOUT_SOFT: Duration = Duration::from_secs(10);
-
 const DOCKER_PROCESS_TIMEOUT_HARD: Duration = Duration::from_secs(12);
-
-const DOCKER_CONTAINER_NAME: &str = "ink-backend";
-
-const DOCKER_WORKDIR: &str = "/builds/contract/";
-
-const DOCKER_OUTPUT: &str = "/playground-result";
 
 // -------------------------------------------------------------------------------------------------
 // TRAIT IMPLEMENTATION
@@ -169,7 +163,7 @@ impl Sandbox {
     pub fn compile(&self, req: &CompilationRequest) -> Result<CompilationResult> {
         self.write_source_code(&req.source)?;
 
-        let command = self.build_compile_command();
+        let command = build_compile_command(&self.input_file, &self.output_dir);
 
         println!("Executing command: \n{:?}", command);
 
@@ -226,99 +220,11 @@ impl Sandbox {
         );
         Ok(())
     }
-
-    fn build_compile_command(&self) -> Command {
-        let mut cmd = self.build_docker_command();
-
-        let execution_cmd = build_execution_command();
-
-        cmd.arg(&DOCKER_CONTAINER_NAME).args(&execution_cmd);
-
-        log::debug!("Compilation command is {:?}", cmd);
-
-        cmd
-    }
-
-    fn build_docker_command(&self) -> Command {
-        let file_name = "lib.rs";
-
-        let mut mount_input_file = self.input_file.as_os_str().to_os_string();
-        mount_input_file.push(":");
-        mount_input_file.push(DOCKER_WORKDIR);
-        mount_input_file.push(file_name);
-
-        let mut mount_output_dir = self.output_dir.as_os_str().to_os_string();
-        mount_output_dir.push(":");
-        mount_output_dir.push(DOCKER_OUTPUT);
-
-        let mut cmd = build_basic_secure_docker_command();
-
-        cmd.arg("--volume")
-            .arg(&mount_input_file)
-            .arg("--volume")
-            .arg(&mount_output_dir);
-
-        cmd
-    }
-}
-// -------------------------------------------------------------------------------------------------
-// MACROS
-// -------------------------------------------------------------------------------------------------
-
-macro_rules! docker_command {
-    ($($arg:expr),* $(,)?) => ({
-        let mut cmd = Command::new("docker");
-        $( cmd.arg($arg); )*
-        cmd
-    });
 }
 
 // -------------------------------------------------------------------------------------------------
 // UTIL FUNCTIONS
 // -------------------------------------------------------------------------------------------------
-
-fn build_basic_secure_docker_command() -> Command {
-    let mut cmd = docker_command!(
-        "run",
-        // "--detach",
-        // "--cap-drop=ALL",
-        // Needed to allow overwriting the file
-        // "--cap-add=DAC_OVERRIDE",
-        //  "--security-opt=no-new-privileges",
-        "--workdir",
-        DOCKER_WORKDIR,
-        // "--net",
-        // "none",
-        "--memory",
-        "512m",
-        "--memory-swap",
-        "640m",
-        "--env",
-        format!(
-            "PLAYGROUND_TIMEOUT={}",
-            DOCKER_PROCESS_TIMEOUT_SOFT.as_secs()
-        ),
-    );
-
-    if cfg!(feature = "fork-bomb-prevention") {
-        cmd.args(&["--pids-limit", "512"]);
-    }
-
-    cmd.kill_on_drop(true);
-
-    cmd
-}
-
-fn build_execution_command() -> Vec<String> {
-    let command = format!(
-        "cargo contract build && mv /target/ink/contract.contract {}",
-        DOCKER_OUTPUT
-    );
-
-    let cmd = vec!["/bin/bash".to_string(), "-c".to_string(), command];
-
-    cmd
-}
 
 fn read(path: &Path) -> Result<Option<Vec<u8>>> {
     let f = match File::open(path) {
